+++ symfem/core/functionals.py
