+++ symfem/core/polynomials.py
