+++ symfem/core/symbolic.py
