+++ symfem/elements/nedelec.py
@@ -1,0 +1,45 @@
+"""Nedelec elements on simplices."""
+
+from ..core.finite_element import FiniteElement, make_integral_moment_dofs
+from ..core.polynomials import polynomial_set, Hcurl_polynomials
+from ..core.functionals import TangentIntegralMoment, IntegralMoment
+from .lagrange import DiscontinuousLagrange, VectorDiscontinuousLagrange
+from .rt import RaviartThomas
+
+
+class NedelecFirstKind(FiniteElement):
+    """Nedelec first kind Hcurl finite element."""
+
+    def __init__(self, reference, order):
+        poly = polynomial_set(reference.tdim, reference.tdim, order - 1)
+        poly += Hcurl_polynomials(reference.tdim, reference.tdim, order)
+        dofs = make_integral_moment_dofs(
+            reference,
+            edges=(TangentIntegralMoment, DiscontinuousLagrange, order - 1),
+            faces=(IntegralMoment, VectorDiscontinuousLagrange, order - 2),
+            volumes=(IntegralMoment, VectorDiscontinuousLagrange, order - 3),
+        )
+
+        super().__init__(reference, poly, dofs, reference.tdim, reference.tdim)
+
+    names = ["Nedelec", "Nedelec1", "N1curl"]
+    min_order = 1
+
+
+class NedelecSecondKind(FiniteElement):
+    """Nedelec second kind Hcurl finite element."""
+
+    def __init__(self, reference, order):
+        poly = polynomial_set(reference.tdim, reference.tdim, order)
+
+        dofs = make_integral_moment_dofs(
+            reference,
+            edges=(TangentIntegralMoment, DiscontinuousLagrange, order),
+            faces=(IntegralMoment, RaviartThomas, order - 1),
+            volumes=(IntegralMoment, RaviartThomas, order - 2),
+        )
+
+        super().__init__(reference, poly, dofs, reference.tdim, reference.tdim)
+
+    names = ["Nedelec2", "N2curl"]
+    min_order = 1
