+++ symfem/elements/rt.py
@@ -1,0 +1,25 @@
+"""Raviart-Thomas elements on simplices."""
+
+from ..core.finite_element import FiniteElement, make_integral_moment_dofs
+from ..core.polynomials import polynomial_set, Hdiv_polynomials
+from ..core.functionals import NormalIntegralMoment, IntegralMoment
+from .lagrange import DiscontinuousLagrange, VectorDiscontinuousLagrange
+
+
+class RaviartThomas(FiniteElement):
+    """Raviart-Thomas Hdiv finite element."""
+
+    def __init__(self, reference, order):
+        poly = polynomial_set(reference.tdim, reference.tdim, order - 1)
+        poly += Hdiv_polynomials(reference.tdim, reference.tdim, order)
+
+        dofs = make_integral_moment_dofs(
+            reference,
+            facets=(NormalIntegralMoment, DiscontinuousLagrange, order - 1),
+            cells=(IntegralMoment, VectorDiscontinuousLagrange, order - 2),
+        )
+
+        super().__init__(reference, poly, dofs, reference.tdim, reference.tdim)
+
+    names = ["Raviart-Thomas", "RT", "N1div"]
+    min_order = 1
